+++ resources/js/bootstrap-table.js
@@ -1,10 +1,3 @@
-// jquery
-import jQuery from "jquery";
-window.$ = jQuery;
-window.jQuery = jQuery;
-
 // bootstrap-table
 import * as bootstrapTable from "bootstrap-table";
 window.bootstrapTable = bootstrapTable
-
-// console.log($);